Compress JP2s in memory
This was compressing JP2s in what happens to be a samba mount, and our
samba provider has become significantly less reliable in the past few
months, causing many apparently nondeterministic errors.

This change is more of a workaround than a real fix, and it's kind of
intentionally hacky. This is the change I made in production in order to
get it working, so I'm pushing it to main and figuring out how to deploy
it from there.

It introduces a risk of filling up memory. For JP2s in the Compressor
class, I worked around it by introducing a new on-disk temporary file so
that /tmp can be cleaned up after each image is compressed, instead of
at the very end.