Hard code NFS space for transient image processing
/tmp turned out to be too small for all but the smallest
batches. A temporary symlink to transient space on quod-prep NFS
proved sufficient and did not reintroduce the CIFS-based
IO issues that led to the original change. This path
would be better managed as a runtime paramater, but absent
developer capacity this is the simplest change to get the
workflow going again for DCU.